Moved code in SchemaInput file
