+++ src/renderer/components/node/NodeInputs.tsx
@@ -1,120 +1,6 @@
-import { NeverType, Type } from '@chainner/navi';
-import { memo, useCallback } from 'react';
-import { useContext, useContextSelector } from 'use-context-selector';
-import {
-    Input,
-    InputData,
-    InputKind,
-    InputSize,
-    NodeSchema,
-    SchemaId,
-} from '../../../common/common-types';
-import { BackendContext } from '../../contexts/BackendContext';
-import { GlobalContext, GlobalVolatileContext } from '../../contexts/GlobalNodeState';
-import { DirectoryInput } from '../inputs/DirectoryInput';
-import { DropDownInput } from '../inputs/DropDownInput';
-import { FileInput } from '../inputs/FileInput';
-import { GenericInput } from '../inputs/GenericInput';
-import { HandleWrapper, InputContainer } from '../inputs/InputContainer';
-import { NumberInput } from '../inputs/NumberInput';
-import { InputProps } from '../inputs/props';
-import { SliderInput } from '../inputs/SliderInput';
-import { TextAreaInput } from '../inputs/TextAreaInput';
-import { TextInput } from '../inputs/TextInput';
-
-const InputComponents: {
-    readonly [K in InputKind]: React.MemoExoticComponent<
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        (props: InputProps<K, any>) => JSX.Element
-    >;
-} = {
-    file: FileInput,
-    directory: DirectoryInput,
-    'text-line': TextInput,
-    text: TextAreaInput,
-    dropdown: DropDownInput,
-    number: NumberInput,
-    slider: SliderInput,
-    generic: GenericInput,
-};
-
-interface SingleInputProps {
-    input: Input;
-    schemaId: SchemaId;
-    nodeId: string;
-    isLocked: boolean;
-    inputData: InputData;
-    inputSize: InputSize | undefined;
-}
-const SingleInput = memo(
-    ({ input, schemaId, nodeId, isLocked, inputData, inputSize }: SingleInputProps) => {
-        const { id: inputId, kind, hasHandle, optional, label } = input;
-
-        const { useInputData, useInputSize: useInputSizeContext } = useContext(GlobalContext);
-        const definitionType = useContextSelector(BackendContext, (c) =>
-            c.functionDefinitions.get(schemaId)?.inputDefaults.get(inputId)
-        );
-
-        const [value, setValue, resetValue] = useInputData(nodeId, inputId, inputData);
-
-        const useInputLocked = useCallback((): boolean => {
-            // TODO: move the function call into the selector
-            // eslint-disable-next-line react-hooks/rules-of-hooks
-            return useContextSelector(GlobalVolatileContext, (c) => c.isNodeInputLocked)(
-                nodeId,
-                inputId
-            );
-        }, [nodeId, inputId]);
-        const useInputType = useCallback((): Type => {
-            // eslint-disable-next-line react-hooks/rules-of-hooks
-            return useContextSelector(GlobalVolatileContext, (c) => {
-                const type = c.typeState.functions.get(nodeId)?.inputs.get(inputId);
-                return type ?? NeverType.instance;
-            });
-        }, [nodeId, inputId]);
-        const useInputSize = useCallback(
-            // eslint-disable-next-line react-hooks/rules-of-hooks
-            () => useInputSizeContext(nodeId, inputId, inputSize),
-            [useInputSizeContext, nodeId, inputId, inputSize]
-        );
-
-        const InputType = InputComponents[kind];
-        const inputElement = (
-            <InputType
-                definitionType={definitionType!}
-                input={input as never}
-                inputKey={`${schemaId}-${inputId}`}
-                isLocked={isLocked}
-                resetValue={resetValue}
-                setValue={setValue}
-                useInputLocked={useInputLocked}
-                useInputSize={useInputSize}
-                useInputType={useInputType}
-                value={value}
-            />
-        );
-
-        return (
-            <InputContainer
-                generic={kind === 'generic'}
-                label={label}
-                optional={optional}
-            >
-                {hasHandle ? (
-                    <HandleWrapper
-                        definitionType={definitionType!}
-                        id={nodeId}
-                        inputId={inputId}
-                    >
-                        {inputElement}
-                    </HandleWrapper>
-                ) : (
-                    inputElement
-                )}
-            </InputContainer>
-        );
-    }
-);
+import { memo } from 'react';
+import { InputData, InputSize, NodeSchema } from '../../../common/common-types';
+import { SchemaInput } from '../inputs/SchemaInput';
 
 interface NodeInputsProps {
     schema: NodeSchema;
@@ -131,7 +17,7 @@ export const NodeInputs = memo(
         return (
             <>
                 {inputs.map((input) => (
-                    <SingleInput
+                    <SchemaInput
                         input={input}
                         inputData={inputData}
                         inputSize={inputSize}
